pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "benchkit"
version = "0.1.0"
description = "Kafka vs Sidekiq benchmark implementation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "confluent-kafka>=2.4.0",
  "jsonschema>=4.22.0",
  "matplotlib>=3.8.0",
  "PyYAML>=6.0.1",
  "redis>=5.0.4",
  "requests>=2.32.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.2.0",
]

[tool.setuptools.packages.find]
include = ["benchkit*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
